Fixed Updating UI from a separate thread.
The issue is that an UI Element are restricted to a single thread the UI Thread. In order update UI elements from a separate thread you have to use a "Handler" to communicate messages between threads. This passes an object and a state back to an object running on the UI Thread, which can then update the UI elements.

The problem was fixed by adapting a training exercise on the android developer site: https://developer.android.com/training/multiple-threads/communicate-ui.html